refactor: keep all content inside form
